fix: clean up and use query further down component tree
